matchmaker/jodi/requirements.txt
python-telegram-bot>=20.0
psycopg2-binary>=2.9
anthropic>=0.18
python-dotenv>=1.0
supabase>=2.0
aiohttp>=3.8.0
python-dateutil>=2.8.0
